Remove BaseUserRole class
